refactor: simplify build and packaging of alert component
