+++ .gitignore
@@ -1,1 +1,5 @@
 node_modules/
+coverage/
+*.d.ts
+!global.d.ts
+custom-elements.json
