+++ packages/components/alert/bs-alert.js
@@ -1,3 +1,6 @@
+import { LitElement, html } from "lit-element";
+import { BsAlertCss } from ".";
+
 /**
  * Alert component
  *
@@ -45,35 +48,93 @@
  * @cssproperty --alert-dark-bd-color - The dark context border color. default: #c6c8ca
  */
 export class BsAlert extends LitElement {
-    static get properties(): {
-        dismissable: {
-            type: BooleanConstructor;
-            reflect: boolean;
+    static get properties() {
+        return {
+            dismissable: {
+                type: Boolean,
+                reflect: true
+            }
         };
-    };
-    static get styles(): import("lit-element").CSSResult[];
-    render(): import("lit-element").TemplateResult;
-    dismissable: boolean;
-    firstUpdated(): void;
-    _addAriaRole(): void;
+    }
+
+    static get styles() {
+        return [BsAlertCss];
+    }
+
+    render() {
+        return html`
+            <slot name="heading"></slot>
+            <slot></slot>
+            <slot name="dismiss"></slot>
+        `;
+    }
+
+    constructor() {
+        super();
+        this.dismissable = false;
+    }
+
+    firstUpdated() {
+        this._addAriaRole();
+        this.addEventListener(
+            "close.button.click",
+            async _ => await this.dismiss()
+        );
+    }
+
+    _addAriaRole() {
+        this.setAttribute("role", "alert");
+    }
+
     /**
      * Dismisses an alert
      *
      * @async
      * @return {Promise} A Promise which is resolved when dismissal has finished with transition.
      */
-    dismiss(): Promise<any>;
+    async dismiss() {
+        if (!this.dismissable) {
+            return;
+        }
+
+        return new Promise(resolve => {
+            const afterTransitionEnd = _ => {
+                this._fireClosedEvent();
+                this.removeEventListener("transitionend", afterTransitionEnd);
+                resolve();
+            };
+
+            this._fireCloseEvent();
+            this.addEventListener("transitionend", afterTransitionEnd);
+            requestAnimationFrame(() => (this.style.opacity = "0"));
+        });
+    }
+
     /**
      * Schedule updates to occur just before the next frame
      */
-    performUpdate(): Promise<void>;
-    _fireCloseEvent(): void;
-    _fireClosedEvent(): void;
-}
-import { LitElement } from "lit-element";
+    async performUpdate() {
+        await new Promise(resolve => requestAnimationFrame(() => resolve()));
+        super.performUpdate();
+    }
 
-declare global {
-    interface HTMLElementTagNameMap {
-      "bs-alert": BsAlert
+    _fireCloseEvent() {
+        const alertCloseEvent = new CustomEvent("close.bs.alert", {
+            bubbles: true,
+            composed: true
+        });
+
+        this.dispatchEvent(alertCloseEvent);
+    }
+
+    _fireClosedEvent() {
+        const alertClosedEvent = new CustomEvent("closed.bs.alert", {
+            bubbles: true,
+            composed: true
+        });
+
+        this.dispatchEvent(alertClosedEvent);
     }
 }
+
+if (!customElements.get("bs-alert")) customElements.define("bs-alert", BsAlert);
