+++ packages/components/alert/index.js
@@ -1,0 +1,5 @@
+export * from './css'
+export { BsAlert } from "./bs-alert.js";
+export { BsAlertHr } from "./bs-alert-hr.js";
+export { BsAlertLink } from "./bs-alert-link.js";
+export { BsAlertDismiss } from "./bs-alert-dismiss.js";
