--- packages/components/alert/types/bs-alert-dismiss.css.d.ts
@@ -1,1 +1,0 @@
-export const BsAlertDismissCss: import("lit-element").CSSResult;
