--- packages/components/alert/types/bs-alert-hr.css.d.ts
@@ -1,1 +1,0 @@
-export const BsAlertHrCss: import("lit-element").CSSResult;
