--- packages/components/alert/types/bs-alert-link.css.d.ts
@@ -1,1 +1,0 @@
-export const BsAlertLinkCss: import("lit-element").CSSResult;
