--- packages/components/alert/types/bs-alert.css.d.ts
@@ -1,1 +1,0 @@
-export const BsAlertCss: import("lit-element").CSSResult;
