--- packages/components/alert/types/index.d.ts
@@ -1,8 +1,0 @@
-export { BsAlert } from "./bs-alert.js";
-export { BsAlertCss } from "./bs-alert.css.js";
-export { BsAlertHr } from "./bs-alert-hr.js";
-export { BsAlertHrCss } from "./bs-alert-hr.css.js";
-export { BsAlertLink } from "./bs-alert-link.js";
-export { BsAlertLinkCss } from "./bs-alert-link.css.js";
-export { BsAlertDismiss } from "./bs-alert-dismiss.js";
-export { BsAlertDismissCss } from "./bs-alert-dismiss.css.js";
